csrf

+++ php/delete_course.php
@@ -12,6 +12,12 @@ if (!isset($_SESSION['username']) || $_SESSION['role'] != 3) { // Only Admin can
     exit();
 }
 
+
+if (!isset($_POST['token']) || $_POST['token'] !== $_SESSION['token']) {
+    header($_SERVER['SERVER_PROTOCOL'] . ' 405 Method Not Allowed');
+    exit;
+} 
+
 // Delete course
 $id = $_GET['id'];
 $sql = "DELETE FROM course WHERE ID = ?";
